requirements.txt
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
transformers==4.37.2
accelerate==1.2.1

numpy==1.26.4
pandas==2.2.3
scipy==1.15.0
scikit-learn==1.6.0
scikit-image==0.25.2

opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
pillow==10.4.0
kornia==0.7.4
basicsr==1.4.2
timm==0.9.6

librosa==0.10.2.post1
soundfile==0.13.0
audioread==3.0.1
decord==0.6.0

bitsandbytes==0.45.0
peft==0.14.0
deepspeed==0.16.2
flash-attn==2.7.1.post4

pyiqa==0.1.13

tqdm==4.67.1
matplotlib==3.10.0
seaborn==0.13.2
einops==0.8.0
omegaconf==2.3.0
hydra-core==1.3.2
colorama==0.4.6
imageio==2.36.1
easydict==1.13
GPUtil==1.4.0

h5py==3.12.1
lmdb==1.6.2
webdataset==0.2.100

requests==2.32.3
fastapi==0.115.6
uvicorn==0.34.0
httpx==0.20.0
openai==0.28.1

jieba==0.42.1
pypinyin==0.53.0
OpenCC==1.1.9

vllm==0.6.6.post1
lmdeploy==0.6.4

wandb==0.15.5

rouge==1.0.1
rouge-chinese==1.0.3
sacrebleu==2.5.1

safetensors==0.5.0
tokenizers==0.15.2
datasets==3.0.1
huggingface-hub==0.25.2
typed-argument-parser==1.10.1
pytz==2024.2